Reland "[Lit][NFC] Refactor shell environment functionality and in-process builtins from TestRunner.py into new modules" (#193759)
Fixed version of #177358 which accidentally removed an import.

This is the first in a series of PRs reworking Lit's handling of
in-process builtins to be run as part of command pipelines, and to
enable users to provide new in-process builtins via Lit's configuration.
The full series of patches can be found here:
https://github.com/BStott6/llvm-project/tree/lit-inproc-builtins.

This work is required for the test daemonization project:
https://discourse.llvm.org/t//88612/9

This PR moves the shell environment functionality and existing
in-process builtins from TestRunner.py into new modules. This is
important not just for organizational purposes but also because,
eventually, user defined in-process builtins will need to be able to
import items from these modules, but these cannot import TestRunner as
this would result in a dependency cycle.
